Optimize badges parsing
Currently we parse badge pages on account becoming free (both game lock and family lock), as well as on new game redeemed on account.

This can be further optimized by avoiding checking badge pages when we can be sure that it didn't change since our last run, so when we already confirmed that we have nothing to farm AND we didn't get disconnect in between. This would make it possible to avoid stupid checks e.g. when user is exiting game being played if we confirmed that there is nothing to farm and we didn't get new game being redeemed event.

If by any chance user will redeem new game while playing something, this will still work correctly and trigger badge check after he's done playing.