Improve fetching master
Previously if user set his own account as master then !loot returned that there is no user with master permission defined, which is correct since definining your own account as master for !loot purposes is invalid.

However, we can slightly improve it, by simply putting those invalid masters as last on the list. This way if there are valid masters, they'll be first, and if not, then the code that already exists in ResponseLoot() will properly return that you can't loot yourself (instead of no master defined response).