Avoid supplying login keys in full-details scenario
Apparently Steam network is too stupid to find out that despite of our invalid login key, we also include password and 2FA token, which are valid. Therefore in order to save potential login failure due to invalid login key, let's not use it at all if it's not needed.

We still save it as we might need to use it in the future.