Avoid matching the same set with the same user as well
In several cases ASF sent more than 1 trade considering the same items, this happens when we have more possible matches than just one but for the same card of other side.

Since we do not preserve the state (due to various reasons), the proper thing to do is to make set exclusive not only to user, but also to a single trade offer. At worst we'll just finish remaining part of it in the next round, if by any chance our trade offer is so huge it stumbles upon 255 items limit.