Slightly improve use case in #1185
API key can be pre-fetched after logging in, and cached for longer than an hour, since bot disconnect already triggers the reset. This way we can avoid asking for that later on, and improve performance of API-key oriented features, without drawbacks from doing so early on, and additionally being covered by LoginLimiterDelay which (accidentally) should help in regards to unsupported amount of bots as well.