Closes #1377
Further improvement in regards to #1375.

Since we have the valid logic for fair bots matching now, we can make use of it for trying to optimize the selection algorithm.

MatchEverything bots should still take priority over fair bots as long as possible, and this is why we'll need **consecutive** 20 empty matches with them before switching to fair bots next. This has the advantage of not switching to fair bots as long as it makes sense, and make the switch only at the point where matching any bots doesn't have a point anymore, and fair bots, despite their lower success chance, are preferred.

While the algorithm is not perfect, it should be good enough for achieving satisfying results, as the next step would involve per-user sets database, which is too much effort for this feature to consider.