Remove ASF-WebConfigGenerator as git submodule
We still use the project and it's not going anywhere, but it's no longer required to be linked directly in ASF core repo. Previously it was linked for translations process, but since it's standalone now, nothing in ASF core references it anymore.

We still need ASF-ui and wiki though, first one as dependency during ASF build process, and second one as we're aiding translation process for that one.