Remove remaining tools
Originally NH2 and NHA2 were added to the repo in early days because I was too lazy to compile those myself and needed to work with them on daily basis.

Since SK2 provides up-to-date binaries now through their CIs, it clutters the repo for no good reason, as everybody who might want to make use of those tools (including me), can just download the official binaries directly.

No reason for this to stay in the repo, as it's not used for any deployment step.