Don't report unknown command on lack of response
Lack of response is actually access denied, not unknown command. Normally users don't get to this point because of !feedback, but in rare case when user has SOME permission, but insufficient one (like FamilySharing for !addlicense), he should get access denied response instead.

Also remove false logging of null response there, it's to be expected in this case.