Don't stack threads in OnHeartBeat()
OnHeartBeat() is the least important function that is called every minute if required, not being able to enter the semaphore means some other action is already in progress, and while it makes sense for stuff like OnPersonaState() to wait for it, because it requires that, heartbeat can totally miss being executed, especially if the previous call didn't manage to finish yet.

It's not destructive as it is right now because tasks would be cancelled eventually when the server starts responding (due to if being removed right now), but it's entirely unnecessary burden.