Optimize HandleCallbacks() routine
We still need the semaphore to ensure we don't launch more than 1 task concurrently, but in unlikely case if we did, it'll just return on the initial call before the second one will finish, as we set KeepRunning = true before spawning a thread.

I don't see a reason why we'd need to enter semaphore on each loop, maybe I forgot about something, but it looks like Archi from the past just didn't notice that.