Slightly improve OnDisconnected() flow
We've added possible delay initializing log on result, but this method is very time-sensitive and there is a possibility of race conditions if we don't reset other stuff, such as AWH, in timely manner ASAP.

It makes sense to move waiting for log on result in a place that isn't so time-sensitive anymore.

This can help with some nasty networking issues where AWH state would get e.g. disabled after already being initialized due to reconnection