Implement extra bullet-proofing and synchronization over connect/disconnect logic
Address @ezhevita findings in regards to race conditions:

- NRE in this lambda function: https://github.com/JustArchiNET/ArchiSteamFarm/blob/1a9f2a23c45f35a969c4c9ec77a4eaa28a7136a3/ArchiSteamFarm/Steam/Bot.cs#L1962
- NRE in the ArchiSteamFarm.Steam.Bot.StopHandlingCallbacks (probably race condition?)

In general, both are caused by race conditions which can happen if user attempts to start/stop bot while critical section of handling callbacks loop is going. The code is overly complex unfortunately, so debugging it/guarantee of safety is problematic.

This commit therefore attempts to fix the underlying issue by synchronizing the code that starts/stops the underlying callbacks handling loop. While the loop itself is already thread-safe, the code that starts/stops it was not before. Now Start() as well as Stop() can not occur concurrently. On top of that, the only other place which has potential to stop the loop - final disconnect, is also guarded with additional condition that it can fire only if we're NOT set to KeepRunning at the time of calling, which should fix the situation where late disconnected callback could potentially stop already triggered new loop.

As usual in such complex situations, time will tell if this fixes all the issues we have.