Update and rename ArchiSteamFarm.container to ArchiSteamFarm@.container
Pending testing.

Mounting container `/app` on `$HOME/ArchiSteamFarm`. The permissions *may* suffice, unless it takes issue when writing outside of `/app` inside of the container because of `ReadWritePaths`.

If systemd prevents writing there AND the container writes anywhere outside of `/app`, `ReadWritePaths` needs to be adjusted accordingly. Without mounting container root itself, I'm not currently aware of how to ascertain that all paths inside the container can also be written before the container is created as the storage location is based on a randomized container UUID.